Change nav buttons to be at top level component
Fix issue

Fix lint issues
