+++ src/js/edu-benefits/containers/BenefitsSelection.jsx
@@ -3,26 +3,15 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import BenefitsSelectionFields from '../components/BenefitsSelectionFields';
-import NavButtons from '../components/NavButtons';
-import { isValidSection } from '../utils/validations';
-import { withRouter } from 'react-router';
-import { veteranUpdateField, ensureFieldsInitialized, updateCompletedStatus } from '../actions/index';
+import { veteranUpdateField } from '../actions/index';
 
 class BenefitsSelection extends React.Component {
   render() {
-    const { section, currentLocation, data, submission, router, onStateChange, dirtyFields, setComplete } = this.props;
-    const navigateTo = path => router.push(path);
+    const { section, data, onStateChange } = this.props;
 
     return (
       <div className="form-panel">
         <BenefitsSelectionFields data={data} section={section} onStateChange={onStateChange}/>
-        <NavButtons
-            submission={submission}
-            path={currentLocation}
-            isValid={isValidSection(currentLocation, data)}
-            dirtyFields={dirtyFields}
-            onNavigate={navigateTo}
-            onComplete={setComplete}/>
       </div>
     );
   }
@@ -32,9 +21,6 @@ function mapStateToProps(state, ownProps) {
   return {
     data: state.veteran,
     section: state.uiState.sections[ownProps.location.pathname],
-    submission: state.uiState.submission,
-    currentLocation: ownProps.location.pathname,
-    router: ownProps.router
   };
 }
 
@@ -42,15 +28,9 @@ function mapDispatchToProps(dispatch) {
   return {
     onStateChange(field, update) {
       dispatch(veteranUpdateField(field, update));
-    },
-    dirtyFields(section) {
-      dispatch(ensureFieldsInitialized(section));
-    },
-    setComplete(section) {
-      dispatch(updateCompletedStatus(section));
     }
   };
 }
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(BenefitsSelection));
+export default connect(mapStateToProps, mapDispatchToProps)(BenefitsSelection);
 export { BenefitsSelection };
