+++ src/js/edu-benefits/containers/EduBenefitsApp.jsx
@@ -3,14 +3,23 @@ import _ from 'lodash';
 
 import { connect } from 'react-redux';
 
+import { withRouter } from 'react-router';
+
 import Nav from '../components/Nav';
+import NavButtons from '../components/NavButtons';
 
 import PerfPanel from '../components/debug/PerfPanel';
 import RoutesDropdown from '../components/debug/RoutesDropdown';
 
+import { isValidSection } from '../utils/validations';
+import { ensureFieldsInitialized, updateCompletedStatus } from '../actions/index';
 
 class EduBenefitsApp extends React.Component {
   render() {
+    const { panels, sections } = this.props.uiState;
+    const { currentLocation, data, submission, router, dirtyFields, setComplete } = this.props;
+    const navigateTo = path => router.push(path);
+
     let devPanel = undefined;
     if (__BUILDTYPE__ === 'development') {
       const queryParams = _.fromPairs(
@@ -25,9 +34,6 @@ class EduBenefitsApp extends React.Component {
       }
     }
 
-    const { panels, sections } = this.props.uiState;
-    const currentLocation = this.props.currentLocation;
-
     return (
       <div className="row">
         {devPanel}
@@ -37,6 +43,13 @@ class EduBenefitsApp extends React.Component {
         <div className="medium-8 columns">
           <div className="progress-box">
             {this.props.children}
+            <NavButtons
+                submission={submission}
+                path={currentLocation.pathname}
+                isValid={isValidSection(currentLocation.pathname, data)}
+                dirtyFields={dirtyFields}
+                onNavigate={navigateTo}
+                onComplete={setComplete}/>
           </div>
         </div>
         <span className="js-test-location hidden" data-location={currentLocation.pathname} hidden></span>
@@ -48,14 +61,24 @@ class EduBenefitsApp extends React.Component {
 function mapStateToProps(state, ownProps) {
   return {
     uiState: state.uiState,
-    currentLocation: ownProps.location
+    currentLocation: ownProps.location,
+    data: state.veteran,
+    submission: state.uiState.submission,
+    router: ownProps.router
   };
 }
 
 // Fill this in when we start using actions
-function mapDispatchToProps() {
-  return {};
+function mapDispatchToProps(dispatch) {
+  return {
+    dirtyFields(section) {
+      dispatch(ensureFieldsInitialized(section));
+    },
+    setComplete(section) {
+      dispatch(updateCompletedStatus(section));
+    }
+  };
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(EduBenefitsApp);
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(EduBenefitsApp));
 export { EduBenefitsApp };
